+++ backend/module-internal/event-webhook/src/main/java/com/toonpick/internal/webhook/slack/config/WebhookConfig.java
@@ -1,4 +1,4 @@
-package com.toonpick.internal.webhook.slack;
+package com.toonpick.internal.webhook.slack.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
