+++ backend/module-internal/event-webhook/src/main/java/com/toonpick/internal/webhook/slack/notifier/SlackNotifier.java
@@ -1,5 +1,6 @@
-package com.toonpick.internal.webhook.slack;
+package com.toonpick.internal.webhook.slack.notifier;
 
+import com.toonpick.internal.webhook.slack.dto.SlackMessage;
 import com.toonpick.type.ErrorCode;
 import lombok.RequiredArgsConstructor;
 import org.slf4j.Logger;
